change: always building new model. fix: clearing tf session between generations.
lesson learned: much better performance + lower memory usage when clearing the session in between generations.
We only need the weights for mutation/crossover processes, so no need to keep referencing the model in those functions.
This makes it possible to clear the session (a copy of the weights after init/mutate/crossover is saved in the wrapper neural network object instead).